requirements.txt
openpyxl==3.1.5
pandas==2.2.2
python-dotenv==1.0.1
yagmail==0.15.293
openai==1.51.2